+++ packages/workbench/src/compute/compute-buffer.ts
@@ -116,7 +116,7 @@ export const ComputeBuffer: LiveComponent<ComputeBufferProps> = (props: ComputeB
   if (!(render ?? children)) return yeet(source);
 
   const content = render ? render(source) : children;
-  const view = provide(ComputeContext, source, content);
+  const view = provide(ComputeContext, [source], content);
 
   if (then) return fence(view, () => then(source));
   return view;